When an unopened database is used, open it automatically.
Motivation:
Before this change, a user had to open a database automatically.  This causes problems with the configuration APIs because the generic Store interface does not define open methods because Memory based stores do not require it.  This meant a user would need to create a store and then cast it, which defeates the purpose of a interface to begin with.

There are a few rough edges in the API still, but this hacks off one of the worst.